+++ app/src/ai/skills/dummy_skill_manager.rs
@@ -2,7 +2,7 @@ use ai::skills::{ParsedSkill, SkillProvider, SkillReference};
 use warp_util::local_or_remote_path::LocalOrRemotePath;
 use warpui::{AppContext, Entity, ModelContext, SingletonEntity};
 
-use crate::ai::skills::SkillDescriptor;
+use crate::ai::skills::{SkillDescriptor, SkillPathQuery};
 
 pub struct SkillManager {}
 
@@ -19,12 +19,17 @@ impl SkillManager {
         vec![]
     }
 
-    pub fn skill_by_path(&self, _skill_path: &LocalOrRemotePath) -> Option<&ParsedSkill> {
+    pub fn skill_by_path<P: SkillPathQuery + ?Sized>(
+        &self,
+        _skill_path: &P,
+    ) -> Option<&ParsedSkill> {
         None
     }
-
-    pub fn reference_for_skill_path(&self, skill_path: &LocalOrRemotePath) -> SkillReference {
-        SkillReference::Path(skill_path.clone())
+    pub fn reference_for_skill_path<P: SkillPathQuery + ?Sized>(
+        &self,
+        skill_path: &P,
+    ) -> SkillReference {
+        SkillReference::Path(skill_path.to_skill_location())
     }
 
     pub fn skill_by_reference(&self, _reference: &SkillReference) -> Option<&ParsedSkill> {
